Changes to make Etoile compile with Clang.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile@5956 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Services/User/DictionaryReader/DictConnection.m
@@ -96,7 +96,7 @@
 
 /** To know whether two remote dictionaries are equal we check if they have the
 	the same host. */
-- (unsigned long) hash
+- (NSUInteger) hash
 {
 	return [host hash] ^ port;
 }
